+++ packages/wm/src/common/events/mod.rs
@@ -6,12 +6,10 @@ mod handle_window_hidden;
 mod handle_window_location_changed;
 mod handle_window_minimize_ended;
 mod handle_window_minimized;
-mod handle_window_moved_end;
 mod handle_window_moved_or_resized_end;
 mod handle_window_shown;
 mod handle_window_title_changed;
 mod handle_window_moved_or_resized_start;
-mod handle_window_resized_end;
 
 pub use handle_display_settings_changed::*;
 pub use handle_mouse_move::*;
